Fixed merge conflict files
